fixing LIMS to int32

+++ MontyCarlo/materials/photon/photon.pyx
@@ -90,7 +90,7 @@ for i in range(Imax + 1): #every possible value of the hash, index = hash
 
     lims.append(np.array([selected[0], selected[-1] , n], dtype = int))
 
-cdef int[:, ::1] LIMS = np.array(lims[1:], dtype = int) ### memory view defined in pxd, cdef double[::1] EAX
+cdef int[:, ::1] LIMS = np.array(lims[1:], dtype = np.int32) ### memory view defined in pxd, cdef double[::1] EAX
